spark/utils/requirements.txt
cassandra-driver
requests